requirements.txt
fastapi
uvicorn
sentence-transformers
openai
faiss-cpu
numpy
